requirements.txt
# Core ML/AI libraries for Backpropagation implementation
numpy>=1.24.0
matplotlib>=3.6.0
scipy>=1.10.0
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.12.0

# OpenAI and API integration
openai>=1.3.0
anthropic>=0.7.0

# MCP (Model Context Protocol) server dependencies
mcp>=0.1.0
uvicorn>=0.24.0
fastapi>=0.104.0
websockets>=12.0
aiohttp>=3.9.0

# Development and research tools
jupyter>=1.0.0
jupyterlab>=4.0.0
ipython>=8.0.0
pandas>=2.0.0
seaborn>=0.12.0

# LaTeX and documentation
pygments>=2.16.0
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Networking and web capabilities
requests>=2.31.0
httpx>=0.25.0
beautifulsoup4>=4.12.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.1
tqdm>=4.66.0
rich>=13.0.0 